+++ cmd/assistantRun.go
@@ -37,7 +37,7 @@ var assistantRunCmd = &cobra.Command{
 		defer xviper.RunMinutes().Done()
 		account := operations.AccountByName(AccountName())
 		if account == nil {
-			pretty.Exit(1, "Could not find account by name: %v", AccountName())
+			pretty.Exit(1, "Could not find account by name: %q", AccountName())
 		}
 		client, err := cloud.NewClient(account.Endpoint)
 		if err != nil {