+++ cmd/download.go
@@ -19,7 +19,7 @@ var downloadCmd = &cobra.Command{
 		}
 		account := operations.AccountByName(AccountName())
 		if account == nil {
-			pretty.Exit(1, "Could not find account by name: %v", AccountName())
+			pretty.Exit(1, "Could not find account by name: %q", AccountName())
 		}
 		client, err := cloud.NewClient(account.Endpoint)
 		if err != nil {