+++ cmd/userinfo.go
@@ -22,7 +22,7 @@ var userinfoCmd = &cobra.Command{
 		}
 		account := operations.AccountByName(AccountName())
 		if account == nil {
-			pretty.Exit(1, "Error: Could not find account by name: %v", AccountName())
+			pretty.Exit(1, "Error: Could not find account by name: %q", AccountName())
 		}
 		client, err := cloud.NewClient(account.Endpoint)
 		if err != nil {